CS 2413 Project One
This project involved the manipulation and storage of Sparse Row Matrices in C++. It involved creating, multiplying, adding, and transposing matrices using the sparse row format. It is a bit of a low point to start out at, as it does not completely work, and is riddled with memory errors. These are not present in any of the other projects in this repository. I keep this one for documentation purposes and a reminder as to the dangers of not setting enough time aside for a project.