CS 2413 Project Two
This project involves the addition, subtraction, multiplication, division and negation of different input chips. It simulates the operations between specialized computer chips, and outputs their results accordingly. My goals for this one were to not repeat the mistakes that led to my project one being unsatisfactory for me, and I succeeded in doing so.