CS 2413 Project Three
This project implements a templated data structure of a queue. It uses this queue to simulate the processing of a CPU, and it manages this via a string of arithmetic operations that are completed in sequence. This was an exploration of pointers and the consequences of a poorly managed one.