CS 2413 Project Four
An M-Tree, or a multiway search tree, is implemented in this project. It provides efficient and well-managed insertion, deletion, and search operations, as well as a rebuild tree option to keep the tree clean. It takes a value, M, and makes sure that the tree has M - 1 values in each parent node as well as M children as a max for each node.